requirements.txt
pytest==7.4.0
requests
gitpython>=3.1.30
python-jenkins
numpy>=1.21.0
pandas>=1.4.0